requirements.txt
a2wsgi==1.10.10
aiosqlite==0.21.0
alembic==1.16.4
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types==0.7.0
anyio==4.10.0
apache-airflow==3.0.4
apache-airflow-core==3.0.4
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.6.2
apache-airflow-providers-common-sql==1.27.5
apache-airflow-providers-smtp==2.2.0
apache-airflow-providers-standard==1.6.0
apache-airflow-task-sdk==1.0.4
archspec @ file:///croot/archspec_1709217642129/work
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
brotli @ file:///croot/brotli-split_1714483155106/work
cadwyn==5.4.4
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorlog==6.9.0
conda @ file:///croot/conda_1733841818148/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1733258728466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda-package-streaming @ file:///croot/conda-package-streaming_1731366181659/work
cron-descriptor==1.4.5
croniter==6.0.0
cryptography @ file:///croot/cryptography_1732129644106/work
deprecated==1.2.18
dill==0.4.0
distro @ file:///croot/distro_1714488253808/work
dnspython==2.7.0
elastic-transport==9.1.0
elasticsearch==9.1.0
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
ffmpy==0.6.1
flask==3.1.2
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2025.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna @ file:///croot/idna_1714398848350/work
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kafka-python==2.2.15
lazy-object-proxy==1.12.0
libcst==1.8.2
libmambapy @ file:///croot/mamba-split_1732896306512/work/libmambapy
linkify-it-py==2.0.3
lockfile==0.12.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
mashumaro==3.16
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1731364904452/work
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging @ file:///croot/packaging_1720101850331/work
pathspec==0.12.1
pendulum==3.1.0
pip @ file:///croot/pip_1723484598856/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
protobuf==6.32.0
protobuf3-to-dict==0.1.5
psutil==7.0.0
psycopg2-binary==2.9.10
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic-core==2.33.2
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pysocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests @ file:///croot/requests_1730999120400/work
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.0
ruamel-yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel-yaml-clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
sentry-sdk==2.35.0
setproctitle==1.3.6
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==1.4.54
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.2
sqlparse==0.5.3
starlette==0.47.3
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tiktoklive==5.0.8
tqdm @ file:///croot/tqdm_1724853939799/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typer==0.16.1
types-requests==2.32.4.20250809
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uc-micro-py==1.0.3
universal-pathlib==0.2.6
urllib3 @ file:///croot/urllib3_1727769808118/work
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
vadersentiment==3.3.2
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.3
wheel==0.44.0
wirerope==1.0.0
wrapt==1.17.3
zipp==3.23.0
zstandard @ file:///croot/zstandard_1731356346222/work
